refactor(api): split plural resource transport schemas
