+++ src/google_work_agent/api/schemas/attachments/__init__.py
@@ -1,0 +1,6 @@
+"""Stable external attachment transport contracts."""
+
+from .get_attachment import AttachmentDescriptorResponse
+from .stage_attachment import StageAttachmentRequest
+
+__all__ = ["AttachmentDescriptorResponse", "StageAttachmentRequest"]
