+++ src/google_work_agent/api/schemas/attachments/stage_attachment.py
@@ -1,0 +1,11 @@
+"""Stage-attachment wire request."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class StageAttachmentRequest(ApiModel):
+    """Base64-encoded JSON upload body."""
+
+    filename: str
+    mime_type: str
+    data_base64: str
