+++ src/google_work_agent/api/schemas/conversations/__init__.py
@@ -1,0 +1,14 @@
+"""Stable external conversation transport contracts."""
+
+from .create_conversation import CreateConversationRequest, ConversationResponse
+from .get_conversation_history import ConversationHistoryResponse
+from .get_latest_run import LatestConversationRunResponse
+from .list_conversations import ConversationListResponse
+
+__all__ = [
+    "ConversationHistoryResponse",
+    "ConversationListResponse",
+    "ConversationResponse",
+    "CreateConversationRequest",
+    "LatestConversationRunResponse",
+]
