+++ src/google_work_agent/api/schemas/conversations/get_conversation_history.py
@@ -1,0 +1,11 @@
+"""Get-conversation-history wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class ConversationHistoryResponse(ApiModel):
+    conversation: dict[str, object]
+    messages: list[dict[str, object]]
+    runs: list[dict[str, object]]
+    truncated: bool
+    api_contract_version: str
