+++ src/google_work_agent/api/schemas/conversations/get_latest_run.py
@@ -1,0 +1,8 @@
+"""Get-latest-conversation-run wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class LatestConversationRunResponse(ApiModel):
+    run: dict[str, object] | None
+    api_contract_version: str
