+++ src/google_work_agent/api/schemas/conversations/list_conversations.py
@@ -1,0 +1,9 @@
+"""List-conversations wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class ConversationListResponse(ApiModel):
+    items: list[dict[str, object]]
+    next_cursor: str | None
+    api_contract_version: str
