+++ src/google_work_agent/api/schemas/events/__init__.py
@@ -1,0 +1,5 @@
+"""Stable external event transport contracts."""
+
+from .get_events import EventEnvelope
+
+__all__ = ["EventEnvelope"]
