+++ src/google_work_agent/api/schemas/resources/count_resources.py
@@ -1,0 +1,9 @@
+"""Count-resources wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class ResourceCountResponse(ApiModel):
+    source: str
+    total_count: int
+    api_contract_version: str
