+++ src/google_work_agent/api/schemas/resources/get_gmail_resource.py
@@ -1,19 +1,6 @@
-"""Google resource projection API schemas."""
+"""Get-Gmail-resource wire response."""
 
 from google_work_agent.api.schemas.common import ApiModel
-
-
-class ResourceListResponse(ApiModel):
-    source: str
-    items: list[dict[str, object]]
-    next_page_token: str | None
-    api_contract_version: str
-
-
-class ResourceCountResponse(ApiModel):
-    source: str
-    total_count: int
-    api_contract_version: str
 
 
 class GmailAttachmentMetadataResponse(ApiModel):
