+++ src/google_work_agent/api/schemas/settings/create_backup.py
@@ -1,0 +1,8 @@
+"""Create-backup wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class BackupResponse(ApiModel):
+    backup: dict[str, object]
+    api_contract_version: str
