+++ src/google_work_agent/api/schemas/settings/create_restore_plan.py
@@ -1,0 +1,12 @@
+"""Create-restore-plan wire contracts."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class RestorePlanRequest(ApiModel):
+    backup_id: str
+
+
+class RestorePlanResponse(ApiModel):
+    plan: dict[str, object]
+    api_contract_version: str
