+++ src/google_work_agent/api/schemas/settings/get_settings.py
@@ -1,0 +1,8 @@
+"""Get-settings wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class SettingsResponse(ApiModel):
+    settings: dict[str, object]
+    api_contract_version: str
