+++ src/google_work_agent/api/schemas/settings/list_backups.py
@@ -1,0 +1,8 @@
+"""List-backups wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class BackupListResponse(ApiModel):
+    items: list[dict[str, object]]
+    api_contract_version: str
