+++ src/google_work_agent/api/schemas/settings/request_shutdown.py
@@ -1,0 +1,8 @@
+"""Request-shutdown wire response."""
+
+from google_work_agent.api.schemas.common import ApiModel
+
+
+class ShutdownResponse(ApiModel):
+    report: dict[str, object]
+    api_contract_version: str
