+++ tests/unit/api/test_schema_ownership.py
@@ -8,6 +8,10 @@ from google_work_agent.api.schemas.actions import (
     PrepareRetryRequestV2,
     RejectActionRequestV2,
 )
+from google_work_agent.api.schemas.attachments import StageAttachmentRequest
+from google_work_agent.api.schemas.conversations import CreateConversationRequest
+from google_work_agent.api.schemas.events import EventEnvelope
+from google_work_agent.api.schemas.resources import ResourceListResponse
 from google_work_agent.api.schemas.runs import (
     CancelRunRequestV2,
     ConfirmationResponseV1,
@@ -15,6 +19,7 @@ from google_work_agent.api.schemas.runs import (
     ResumeRunRequestV2,
     StartRunRequest,
 )
+from google_work_agent.api.schemas.settings import PatchSettingsRequest
 
 
 def test_action_transport_contracts_live_in_operation_modules() -> None:
@@ -32,7 +37,23 @@ def test_run_transport_contracts_live_in_operation_modules() -> None:
     assert ResolveRecoveryRequestV1.__module__.endswith(".runs.resolve_recovery")
 
 
-def test_broad_action_and_run_schema_modules_are_retired() -> None:
+def test_other_plural_resource_contracts_live_in_operation_modules() -> None:
+    assert StageAttachmentRequest.__module__.endswith(".attachments.stage_attachment")
+    assert CreateConversationRequest.__module__.endswith(".conversations.create_conversation")
+    assert EventEnvelope.__module__.endswith(".events.get_events")
+    assert ResourceListResponse.__module__.endswith(".resources.list_resources")
+    assert PatchSettingsRequest.__module__.endswith(".settings.update_settings")
+
+
+def test_retired_broad_plural_resource_schema_modules_are_absent() -> None:
     api_root = Path(__file__).resolve().parents[3] / "src" / "google_work_agent" / "api" / "schemas"
-    assert not (api_root / "actions.py").exists()
-    assert not (api_root / "runs.py").exists()
+    for filename in (
+        "actions.py",
+        "attachments.py",
+        "conversations.py",
+        "events.py",
+        "resources.py",
+        "runs.py",
+        "settings.py",
+    ):
+        assert not (api_root / filename).exists()
